Randerizando sem mvpMatrix.

+++ animation/Frame.cpp
@@ -1,16 +1,16 @@
 #include "Frame.hpp"
 
-Frame::Frame(float *vertices, int verticeLength, short *drawOrder, int drawOrderLength){
+Frame::Frame(float *vertices, int verticeLength, GLuint *drawOrder, int drawOrderLength){
     this->vertices = vertices;
     this->drawOrder = drawOrder;
     this->drawableBuffer = new DrawableBuffer(vertices, verticeLength, drawOrder, drawOrderLength);
 }
 
-float* Frame::getVertices(){
+GLfloat* Frame::getVertices(){
     return this->vertices;
 }
 
-short* Frame::getDrawOrder(){
+GLuint* Frame::getDrawOrder(){
     return this->drawOrder;
 }
